Safari-611.1.10/LayoutTests/svg/css/shape-rendering-parsing-expected.txt

Tests the parsing of shape-rendering property values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.defaultView.getComputedStyle(group, null).shapeRendering is "auto"
PASS document.defaultView.getComputedStyle(group, null).shapeRendering is "crispedges"
PASS document.defaultView.getComputedStyle(group, null).shapeRendering is "crispedges"
PASS document.defaultView.getComputedStyle(group, null).shapeRendering is "optimizeSpeed"
PASS document.defaultView.getComputedStyle(group, null).shapeRendering is "geometricPrecision"
PASS successfullyParsed is true

TEST COMPLETE
